requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
asgiref==3.4.1
astroid==2.9.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==36.0.1
DBUtils==1.3
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.1
django-redis==5.2.0
djangorestframework==3.12.4
gmssl==3.2.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
oss2==2.15.0
packaging==21.3
pycparser==2.21
pycryptodome==3.13.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
qcloudsms-py==0.1.4
redis==4.1.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tencentcloud-sdk-python==3.0.575
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
Faker==13.6.0
